app/requirements.txt
streamlit
langchain
langchain-openai
chroma
python-dotenv
